pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "takehome"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "minio>=7.2.0",
    "datasets>=2.14.0",
    "pandas>=2.0.0",
    "tqdm>=4.66.0",
    "click>=8.1.0",
    "mlflow>=2.12.0",
    "psycopg2-binary>=2.9.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "greenlet>=3.2.4",
    "boto3>=1.40.21",
    "tiktoken>=0.11.0",
]

[project.scripts]
mlflow-server = "app.scripts.mlflow_server:start_mlflow_server"
db-manager = "app.process.cli.db_manager:cli"
pipeline = "app.process.cli.pipeline:cli"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-benchmark>=4.0.0",
    "memory-profiler>=0.61.0",
]
